backend/requirements.txt
Django==5.0.7
pillow==8.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0


